files (41)/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
jinja2==3.1.4
pandas==2.2.2
openpyxl==3.1.2
numpy==1.26.4
python-dotenv==1.0.1
aiofiles==23.2.1
python-pptx==0.6.23
langchain-openai==0.1.8
langchain-core==0.2.9
langchain-experimental==0.0.61
langchain-community==0.2.5
langgraph==0.1.5
